Create README.md
We are curating a set of arts and related datasets for the Arts Access Datathon (http://artsdatathon.org) happening Saturday, April 22, 2017. This list will continue to grow until the day of the datathon, so come back often to see what we’ve added. Be sure to check out the Data Analysis Tools page (http://artsdatathon.org/data/data-tools/) and get to know some of the newest and some tried-and-true ways to explore data.

Have a dataset to suggest? Contact us at info@artsdatathon.org.