LAPhil Data sets
The LA Phil's Symphonies for School program provides students the opportunity to connect to passionate musicians, a world-class orchestra, and a concert hall that students have described as a silver tulip. These datasets provide a list of schools that are a part of the program and participation data associated with the latest programming.